@@ -20,27 +20,27 @@
 		"@playwright/test": "^1.45.2",
 		"@sveltejs/adapter-auto": "^3.0.0",
 		"@sveltejs/adapter-node": "^2.0.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/kit": "^2.5.27",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@typescript-eslint/eslint-plugin": "^6.21.0",
 		"@typescript-eslint/parser": "^6.21.0",
-		"@zerodevx/svelte-toast": "^0.9.5",
+		"@zerodevx/svelte-toast": "^0.9.6",
 		"autoprefixer": "^10.4.19",
 		"eslint": "^8.57.0",
 		"eslint-config-prettier": "^8.10.0",
-		"eslint-plugin-svelte": "^2.43.0",
-		"monaco-editor": "0.40.0",
-        "@gregofi1/regex-tooling": "^0.0.4",
+		"eslint-plugin-svelte": "^2.45.1",
+		"monaco-editor": "0.52.0",
+		"@gregofi1/regex-tooling": "^0.0.4",
 		"postcss": "^8.4.39",
-		"prettier": "^2.8.8",
-		"prettier-plugin-svelte": "^2.10.1",
-		"svelte": "^4.2.18",
+		"prettier": "^3.1.0",
+		"prettier-plugin-svelte": "^3.2.6",
+		"svelte": "^5.0.0",
 		"svelte-adapter-bun": "^0.5.2",
-		"svelte-check": "^3.8.4",
+		"svelte-check": "^4.0.0",
 		"tailwindcss": "^3.4.6",
 		"tslib": "^2.6.3",
 		"typescript": "^5.5.3",
-		"vite": "^5.0.0",
+		"vite": "^5.4.4",
 		"vitest": "^1.0.0"
 	},
 	"type": "module",
