@@ -1,4 +1,4 @@
 <div
 	class="absolute inset-0 m-auto border-gray-300 h-20 w-20
 animate-spin rounded-full border-8 border-t-blue-600"
-/>
+></div>
